Create JumpGameTwo.cs
This solution iterates through the array, keeping track of the farthest index that can be reached from the current position. When the current position reaches the end of the current jump (i.e., i == currentJumpEnd), it means we need to make another jump, so we update the currentJumpEnd to the farthest reachable index. Finally, we return the total number of jumps needed to reach the end.